pyproject.toml
[project]
name = "search-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.19.0",
    "litellm>=1.80.7",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
